volume: refactor layered asset volumes and centralized checks.
This refactor consolidates volume permission/identity logic and tightens the
layered overlay + embedded-assets model.

Motivation
- Reduce duplicated read/write validation across local/overlay volumes.
- Make overlay/asset volumes consistently identify themselves and their backing.
- Prepare volume and asset code to share common helpers and a unified error
  surface.

Key changes
- Centralized I/O/read/write checks and normalized volume identity/overlap
  labeling across volume implementations.
- Reworked overlay volume interfaces so the layered asset flow is explicit.
- Added shared helpers in `src/volume/Volume_utils.cpp` to avoid scattering
  common logic.
- Extended `MemoryZip` and associated interfaces to better support embedded
  assets and the new asset registry flow.
- Updated ext4/fat32 volume implementations and tests to use the new APIs.
- Updated build + module integration (`meson.build`, `src/module.cpp`, and
  assets wiring) to include the new/changed volume sources.

No functional change is intended beyond making the behavior more consistent and
easier to reason about (especially for overlay + embedded-asset paths).

+++ src/volume/VolumeFile.cpp
@@ -148,7 +148,7 @@ std::unique_ptr<VolumeFile> VolumeFile::getParentFile() const {
     if (!m_volume || m_path.empty()) {
         return nullptr;
     }
-    int lastSlash = m_path.find_last_of('/');
+    size_t lastSlash = m_path.find_last_of('/');
     if (lastSlash == std::string::npos) {
         return nullptr;
     }
